+++ examples/Dataset-001.yaml
@@ -1,0 +1,7 @@
+# Example data object
+---
+entries:
+  - id: example:Dataset001
+    name: foo bar
+    primary_email: foo.bar@example.com
+    age_in_years: 33
