+++ project/jsonld/datalad_schema.context.jsonld
@@ -1,7 +1,7 @@
 {
    "comments": {
       "description": "Auto generated by LinkML jsonld context generator",
-      "generation_date": "2023-11-12T16:13:32",
+      "generation_date": "2023-11-13T11:12:25",
       "source": "datalad_schema.yaml"
    },
    "@context": {
