+++ project/shex/datalad_schema.shex
@@ -2,7 +2,7 @@ BASE <https://w3id.org/psychoinformatics-de/datalad-schema/>
 PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
 PREFIX xsd: <http://www.w3.org/2001/XMLSchema#>
 PREFIX linkml: <https://w3id.org/linkml/>
-PREFIX schema1: <http://schema.org/>
+PREFIX schema: <http://schema.org/>
 
 
 linkml:String xsd:string
@@ -45,9 +45,9 @@ linkml:Sparqlpath xsd:string
 
 <Dataset> CLOSED {
     (  $<Dataset_tes> (  &<NamedThing_tes> ;
-          rdf:type [ schema1:Thing ] ? ;
-          schema1:email @linkml:String ? ;
-          schema1:birthDate @linkml:Date ? ;
+          rdf:type [ schema:Thing ] ? ;
+          schema:email @linkml:String ? ;
+          schema:birthDate @linkml:Date ? ;
           <age_in_years> @linkml:Integer ? ;
           <vital_status> @<PersonStatus> ?
        ) ;
@@ -63,10 +63,10 @@ linkml:Sparqlpath xsd:string
 
 <NamedThing>  (
     CLOSED {
-       (  $<NamedThing_tes> (  schema1:name @linkml:String ? ;
-             schema1:description @linkml:String ?
+       (  $<NamedThing_tes> (  schema:name @linkml:String ? ;
+             schema:description @linkml:String ?
           ) ;
-          rdf:type [ schema1:Thing ]
+          rdf:type [ schema:Thing ]
        )
     } OR @<Dataset>
 )
